pyproject.toml
[tool.poetry]
name = "mutagenesis-visualization"
version = "0.4.0"
description = ""
authors = ["Frank <yfhidalgoruiz@berkeley.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
mypy = "^0.812"
pylint = "^2.6.0"
ipython = "^7.19.0"
numpy = "^1.19.5"
pandas = "^1.2.0"
dataclasses-json = "^0.5.2"
scipy = "^1.5"
matplotlib = "^3.3"
seaborn="0.10.0"
scikit-learn = "^0.24.2"
adjustText = "^0.7.3"
biopython = "^1.79"
freesasa = "^2.1.0"
plotly = "^5.1.0"
openpyxl = "^3.0.7"
ipympl = "^0.7.0"
xlrd = "^2.0.1"
statsmodels = "^0.12.2"
XlsxWriter = "^1.4.4"
yapf = "^0.31.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^5.4.3"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
